requirements.txt
PySide6
PyOpenGL
numpy
Pillow
openai
sounddevice
scipy
typecast-python
httpx
openai-whisper
# live2d-py  <-- Uncomment if you have the prebuilt wheels or can build it. 
# Otherwise, the app will run in "Mock" mode.
